refactor: elevate general types; renames
