+++ src/chart_types/partition_chart/layout/viewmodel/link_text_layout.ts
@@ -1,10 +1,11 @@
 import { Distance } from '../types/geometry_types';
 import { Config } from '../types/config_types';
 import { TAU, trueBearingToStandardPositionAngle } from '../utils/math';
-import { LinkLabelVM, ShapeTreeNode, ValueFormatter } from '../types/viewmodel_types';
+import { LinkLabelVM, ShapeTreeNode } from '../types/viewmodel_types';
 import { meanAngle } from '../geometry';
 import { TextMeasure } from '../types/types';
 import { AGGREGATE_KEY } from '../utils/group_by_rollup';
+import { ValueFormatter } from '../../../../utils/commons';
 
 // todo modularize this large function
 export function linkTextLayout(
